requirements.txt
Django==3.1
python-decouple==3.3
djangorestframework==3.11.0
